+++ vyre-libs-parsing/src/parsing/go/parse/structure/mod.rs
@@ -3,6 +3,7 @@
 
 mod decl_span;
 mod declarations;
+mod interface_decl;
 mod packages;
 
 pub use declarations::go_extract_declarations;